pyproject.toml
[project]
name = "2-mda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.136.0",
    "lxml>=6.1.0",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn>=0.44.0",
]
